extracted out cli functions in cli_utils.hpp
